Refactor by implementing serializer
