+++ fakestagram/DataSource/RestClient.swift
@@ -17,12 +17,6 @@ class RestClient<T: Codable> {
         return encoder
     }()
 
-    let decoder: JSONDecoder = {
-        let decoder = JSONDecoder()
-        decoder.keyDecodingStrategy = .convertFromSnakeCase
-        return decoder
-    }()
-
     init(client: Client, basePath: String) {
         self.client = client
         self.basePath = basePath
@@ -31,19 +25,9 @@ class RestClient<T: Codable> {
     // eg. POST /posts/
     func create(_ data: T, success: @escaping (T?) -> Void) {
         let payload = try? encoder.encode(data)
-        client.post(path: basePath, body: payload) { [unowned self] data in
-                   guard let data = data else {
-                       success(nil)
-                       return
-                   }
-                   do {
-                       let json = try self.decoder.decode(T.self, from: data)
-                       success(json)
-                   } catch let err {
-                       print("Invalid JSON format: \(err.localizedDescription)")
-                   }
-               }
-
+        client.post(path: basePath, body: payload) { data in
+            CodableSerializer(data: data).async(result: success)
+        }
     }
 
     // eg. GET /posts/
@@ -53,51 +37,22 @@ class RestClient<T: Codable> {
         if let uid = id {
             path += "/\(uid)"
         }
-        client.get(path: path) { [unowned self] data in
-            guard let data = data else {
-                DispatchQueue.main.async { success(nil) }
-                return
-            }
-            do {
-                let json = try self.decoder.decode(T.self, from: data)
-                DispatchQueue.main.async { success(json) }
-            } catch let err {
-                print("Invalid JSON format: \(err.localizedDescription)")
-            }
+        client.get(path: path) { data in
+            CodableSerializer(data: data).async(result: success)
         }
     }
     // eg. PUT /posts/:id
     func update(id: String, data: T, success: @escaping (T?) -> Void) {
         let payload = try? encoder.encode(data)
-        client.put(path: "\(basePath)/\(id)", body: payload) { [unowned self] data in
-            guard let data = data else {
-                success(nil)
-                return
-            }
-            do {
-                let json = try self.decoder.decode(T.self, from: data)
-                success(json)
-            } catch let err {
-                print("Invalid JSON format: \(err.localizedDescription)")
-            }
+        client.put(path: "\(basePath)/\(id)", body: payload) { data in
+            CodableSerializer(data: data).async(result: success)
         }
     }
 
-
     // eg. DELETE /posts/:id
     func delete(id: String, success: @escaping (T?) -> Void) {
-        client.delete(path: "\(basePath)/\(id)") { [unowned self] data in
-                guard let data = data else {
-                    success(nil)
-                    return
-                }
-                do {
-                    let json = try self.decoder.decode(T.self, from: data)
-                    success(json)
-                } catch let err {
-                    print("Invalid JSON format: \(err.localizedDescription)")
-                }
-            }
+        client.delete(path: "\(basePath)/\(id)") { data in
+            CodableSerializer(data: data).async(result: success)
+        }
     }
-
 }
